Separate tally scoring from Tally object methods
